requirements.txt
langchain>=0.1.0
langchain-community>=0.0.30
langchain-openai>=0.0.8
chromadb>=0.4.24
openai
tqdm
python-dotenv
pdfplumber
fastapi
uvicorn[standard]
chromadb
langchain